merged

+++ server/src/main/java/net/guides/springboot2/crud/controller/BazacalculController.java
@@ -50,7 +50,7 @@ public class BazacalculController {
 
 	@GetMapping("cm/{ida}/mo={luna}&y={an}")
 	public BazaCalculCMDTO getBazaCalculCM(
-		@PathVariable("ida") long ida,
+		@PathVariable("ida") int ida,
 		@PathVariable("luna") int luna,
 		@PathVariable("an") int an) {
 			return bazaCalculService.getBazaCalculCM(luna, an, ida);
